+++ semantic-ui/src/main/scala/react/semanticui/collections/form/FormTextArea.scala
@@ -115,7 +115,7 @@ object FormTextArea {
     var onInput: js.UndefOr[TextArea.RawEvent] = js.native
 
     /** Indicates row count for a TextArea. */
-    var rows: js.UndefOr[Double | String] = js.native
+    var rows: js.UndefOr[Int | String] = js.native
 
     /** The value of the textarea. */
     var value: js.UndefOr[Double | String] = js.native
